thoughty_backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djoser>=2.2.0
PyJWT>=2.6.0
groq>=0.4.0
requests>=2.28.0
dotenv
django-cors-headers
django-extensions